requirements.txt
ray[tune]~=1.7.1
omegaconf~=2.1.1
tqdm~=4.62.3
numpy~=1.21.3
pillow~=8.4.0
gym~=0.21.0
wandb~=0.12.5
git+git://github.com/SpirinEgor/mapgen@master#egg=mapgen

# ray requirements
pandas~=1.3.4
dm_tree~=0.1.6
scikit-image~=0.18.3
lz4~=3.1.3
torch~=1.10.0
